+++ packages/theme/CHANGELOG.md
@@ -2,6 +2,10 @@
 
 ## Unreleased
 
+### New Features
+
+-   Add `generateColorTokens`, which derives color token values from seed colors outside React. Also available as `@wordpress/theme/colors` for consumers that cannot load React ([#82425](https://github.com/WordPress/gutenberg/pull/82425)).
+
 ### Internal
 
 -   Run rendered Theme tests in Vitest Browser Mode ([#80995](https://github.com/WordPress/gutenberg/pull/80995)).